requirements.txt
opencv-python==4.8.1.78
easyocr==1.7.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
numpy==1.24.4
streamlit==1.29.0